refactor: modularize codebase into domain-specific modules
- Extract CLI functionality into src/cli/ directory
- Create domain-specific modules in src/domains/
- Move core algorithms to src/core/
- Reorganize MCP tools into src/mcp/tools/
- Add shared utilities in src/shared/
- Simplify main index.js to use modular architecture
- Update all tests to work with new module structure
- Preserve existing functionality and CLI interface
