+++ test/core.test.js
@@ -4,7 +4,9 @@ import { execSync } from 'node:child_process';
 import fs from 'node:fs';
 import path from 'node:path';
 import os from 'node:os';
-import { parseArguments, calculateFinalAmount, formatOutput } from '../src/core.js';
+import { parseArguments } from '../src/shared/argument-parsing-simple.js';
+import { calculateFinalAmount } from '../src/core/calculation.js';
+import { formatOutput } from '../src/core.js';
 
 const CLI_PATH = path.join(process.cwd(), 'index.js');
 const CONFIG_DIR = path.join(os.homedir(), '.config', 'gift-calc');
